Chendge edit faculty changes and delete bug fix
