+++ CourseAssignmentSystem-master/app/helpers/faculties_helper.rb
@@ -1,0 +1,2 @@
+module FacultiesHelper
+end
