+++ CourseAssignmentSystem-master/app/models/course.rb
@@ -2,4 +2,6 @@ class Course < ActiveRecord::Base
 	has_many :course_assignments, :dependent => :destroy
 	has_many :faculty_courses, :dependent => :destroy
 	has_many :faculty_preferences, :through => :faculty_courses, :dependent => :destroy
+
+	validates :course_name, uniqueness: true
 end
