+++ CourseAssignmentSystem-master/app/models/faculty_preference.rb
@@ -1,5 +1,6 @@
 class FacultyPreference < ActiveRecord::Base
 	belongs_to :faculty_course
+	belongs_to :faculty
 	belongs_to :preference1, class_name: "Preference"
 	belongs_to :preference2, class_name: "Preference"
 	belongs_to :preference3, class_name: "Preference"
