+++ CourseAssignmentSystem-master/spec/controllers/faculties_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe FacultiesController, type: :controller do
+
+end
